Rename initial ai customizations -> aichat settings
