+++ dashboard/app/controllers/levels_controller.rb
@@ -573,7 +573,7 @@ class LevelsController < ApplicationController
     # Parse a few specific JSON fields used by modern (Lab2) labs so that they are
     # stored in the database as a first-order member of the properties JSON, rather
     # than simply as a string of JSON belonging to a single property.
-    [:level_data, :initial_ai_customizations, :validations, :panels].each do |key|
+    [:level_data, :aichat_settings, :validations, :panels].each do |key|
       level_params[key] = JSON.parse(level_params[key]) if level_params[key]
     end
     # Delete validations from level data if present. We'll use the validations in level properties instead.
