+++ dashboard/app/models/levels/aichat.rb
@@ -29,7 +29,7 @@ class Aichat < Level
     system_prompt
     bot_title
     bot_description
-    initial_ai_customizations
+    aichat_settings
   )
 
   def self.create_from_level_builder(params, level_params)
