+++ dashboard/app/views/levels/editors/_aichat.html.haml
@@ -1,3 +1,3 @@
-= render partial: 'levels/editors/fields/ai_customizations', locals: {f: f}
+= render partial: 'levels/editors/fields/aichat_settings', locals: {f: f}
 = render partial: 'levels/editors/fields/instructions_area', locals: {f: f}
 = render partial: 'levels/editors/fields/audit_log', locals: {f: f}
