+++ dashboard/app/views/levels/editors/fields/_aichat_settings.html.haml
@@ -1,0 +1,8 @@
+%h1.control-legend{data: {toggle: "collapse", target: "#aichat-settings-container"}}
+  AI Chat Settings
+
+#aichat-settings-container.in.collapse
+
+- content_for :body_scripts do
+  %script{src: webpack_asset_path('js/levels/editors/fields/_aichat_fields.js'),
+        data: {aichatsettings: (@level.properties["aichat_settings"]).to_json}}
